Support providing extra data processing functions.
- via `--custom-functions [PYTHON_FILE]`
